+++ R/open_dataset.R
@@ -14,7 +14,7 @@
 #'  column name across all files (NOTE: this can add considerably to
 #'  the initial execution time)
 #' @param format The format of the dataset files. One of `"parquet"`, `"csv"`,
-#' `"tsv"`, or `"sf"` (spatial vector files supported by the sf package / GDAL).
+#'  or `"sf"` (spatial vector files supported by the sf package / GDAL).
 #'  if no argument is provided, the function will try to guess the type based
 #'  on minimal heuristics.
 #' @param conn A connection to a database.
@@ -31,6 +31,8 @@
 #' the table name.
 #' @param recursive should we assume recursive path? default TRUE. Set to FALSE
 #' if trying to open a single, un-partitioned file.
+#' @param parser_options additional options passed to the parser, e.g. to
+#' read_csv(), see <https://duckdb.org/docs/stable/data/csv/overview.html#parameters>
 #' @param ... optional additional arguments passed to [duckdb_s3_config()].
 #'   Note these apply after those set by the URI notation and thus may be used
 #'   to override or provide settings not supported in that format.
@@ -55,6 +57,14 @@
 #' efi <- open_dataset("s3://neon4cast-scores/parquet/aquatics",
 #'                     s3_access_key_id="",
 #'                     s3_endpoint="data.ecoforecast.org")
+#' 
+#' # Use parser-options for non-standard csv:
+#'  cars <- tempfile() # dummy data
+#'  write.table(mtcars, cars, row.names = FALSE)
+#' 
+#' # Note nested quotes on parser option for delimiter:
+#'  df <- open_dataset(cars, format = "csv",
+#'                     parser_options = c(delim = "' '", header = TRUE))
 #'
 #' @export
 open_dataset <- function(sources,
@@ -67,6 +77,7 @@ open_dataset <- function(sources,
                          mode = "VIEW",
                          filename = FALSE,
                          recursive = TRUE,
+                         parser_options = list(),
                          ...) {
 
   format <- select_format(sources, format)
@@ -90,7 +101,8 @@ open_dataset <- function(sources,
                              mode = mode,
                              hive_partitioning = hive_style,
                              union_by_name = unify_schemas,
-                             filename = filename
+                             filename = filename,
+                             parser_options = parser_options
                              )
 
   DBI::dbSendQuery(conn, view_query)
@@ -115,7 +127,7 @@ select_format <- function(sources, format) {
   }
 
   if( fs::is_dir(sources) ) {
-    sources <- fs::dir_ls(sources, recurse = TRUE, type="file")
+    sources <- fs::dir_ls(sources, recurse = TRUE, type = "file")
     sources <- sources[[1]]
   }
 
@@ -155,7 +167,8 @@ query_string <- function(tblname,
                          mode = c("VIEW", "TABLE"),
                          hive_partitioning = TRUE,
                          union_by_name = FALSE,
-                         filename = FALSE) {
+                         filename = FALSE, 
+                         parser_options = list()) {
  # format <- match.arg(format)
   scanner <- switch(format,
                     "parquet" = "parquet_scan(",
@@ -172,9 +185,9 @@ query_string <- function(tblname,
          "TABLE" = "TABLE")
 
   tabular_options <- paste0(
-    ", HIVE_PARTITIONING=",hive_partitioning,
-    ", UNION_BY_NAME=",union_by_name,
-    ", FILENAME=",filename)
+    ", HIVE_PARTITIONING=", hive_partitioning,
+    ", UNION_BY_NAME=", union_by_name,
+    ", FILENAME=", filename)
 
   options <- switch(format,
                     "parquet" = tabular_options,
@@ -182,6 +195,13 @@ query_string <- function(tblname,
                     "sf" = "",
                     tabular_options
   )
+
+  # append any custom options
+  if(length(parser_options) > 0) {
+    pairs <- paste(names(parser_options), "=", parser_options,
+                 sep = "", collapse = ", ")
+    options <- paste0(c(options, pairs), collapse=", ")
+  }
   paste0(
     paste("CREATE", mode, tblname, "AS SELECT * FROM "),
     paste0(scanner, source_uris, options,